refactor: move encrypt / decrypt methods without output path to default interface implementation
