+++ src/main/java/ru/obninsk/iate/easycipher/lib/abstractions/ICryptoService.java
@@ -1,5 +1,7 @@
 package ru.obninsk.iate.easycipher.lib.abstractions;
 
+import org.jetbrains.annotations.NotNull;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -7,17 +9,26 @@ import java.nio.file.Path;
 public interface ICryptoService {
     boolean encryptFile(Path filePath, String key, Path out);
 
-    boolean encryptFile(Path filePath, String key);
-
     boolean decryptFile(Path filePath, String key, Path out);
-
-    boolean decryptFile(Path filePath, String key);
 
     boolean encryptDirectory(Path dirPath, String key, Path out);
 
-    boolean encryptDirectory(Path dirPath, String key);
-
     boolean decryptDirectory(Path dirPath, String key, Path out);
+
+    default boolean encryptFile(@NotNull Path filePath, String key) {
+        var newFileName = filePath.getFileName().toString() + ".enc";
+        return encryptFile(filePath, key, filePath.resolveSibling(newFileName));
+    }
+
+    default boolean decryptFile(@NotNull Path filePath, String key) {
+        var newFileName = filePath.getFileName().toString() + ".dec";
+        return decryptFile(filePath, key, filePath.resolveSibling(newFileName));
+    }
+
+    default boolean encryptDirectory(@NotNull Path directoryPath, String key) {
+        var newFileName = directoryPath.getFileName().toString() + ".encd";
+        return encryptDirectory(directoryPath, key, directoryPath.resolveSibling(newFileName));
+    }
 
     default boolean decryptDirectory(Path dirPath, String key) {
         try {
